pyproject.toml
[project]
name = "flight-delay-analytics"
version = "1.0.0"
description = "End-to-end airline operations analytics platform analyzing 12M+ flights to surface actionable delay reduction strategies"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "duckdb>=0.10.0",
    "openpyxl>=3.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.4.0",
    "shap>=0.44.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
